--- custom_components/sems-wallbox/strings.json
@@ -1,71 +1,0 @@
-{
-  "title": "GoodWe SEMS Wallbox API",
-  "exceptions": {
-    "set_charge_mode_failed_fast": {
-      "message": "Switching to Fast charging mode failed (timeout or network error). The previous mode has been restored."
-    },
-    "set_charge_mode_failed_pv_priority": {
-      "message": "Switching to PV Priority mode failed (timeout or network error). The previous mode has been restored."
-    },
-    "set_charge_mode_failed_pv_and_battery": {
-      "message": "Switching to PV & Battery mode failed (timeout or network error). The previous mode has been restored."
-    },
-    "set_charge_power_failed": {
-      "message": "Failed to set charge power to {value} kW. The previous value has been restored."
-    }
-  },
-  "options": {
-    "step": {
-      "init": {
-        "data": {
-          "scan_interval": "Idle update interval (seconds)",
-          "scan_interval_charging": "Charging update interval (seconds)"
-        },
-        "data_description": {
-          "scan_interval": "How often to poll when not charging (10–300 s)",
-          "scan_interval_charging": "How often to poll while actively charging (5–120 s)"
-        }
-      }
-    }
-  },
-  "config": {
-    "step": {
-      "user": {
-        "data": {
-          "username": "[%key:common::config_flow::data::username%]",
-          "password": "[%key:common::config_flow::data::password%]"
-        }
-      },
-      "plant": {
-        "data": {
-          "plant_id": "Plant / station"
-        }
-      },
-      "charger": {
-        "data": {
-          "wallbox_serial_No": "Wallbox"
-        }
-      },
-      "charger_manual": {
-        "data": {
-          "wallbox_serial_No": "Wallbox serial number",
-          "plant_id": "Plant ID (optional)"
-        }
-      },
-      "model": {
-        "data": {
-          "product_model": "Product model"
-        }
-      }
-    },
-    "error": {
-      "cannot_connect": "[%key:common::config_flow::error::cannot_connect%]",
-      "invalid_auth": "[%key:common::config_flow::error::invalid_auth%]",
-      "no_chargers_found": "No EV chargers found. Enter the serial number manually.",
-      "unknown": "[%key:common::config_flow::error::unknown%]"
-    },
-    "abort": {
-      "already_configured": "[%key:common::config_flow::abort::already_configured_device%]"
-    }
-  }
-}
