--- custom_components/sems-wallbox/translations/cs.json
@@ -1,148 +1,0 @@
-{
-    "title": "GoodWe SEMS API pro Wallbox",
-    "exceptions": {
-        "set_charge_mode_failed_fast": {
-            "message": "Přepnutí do režimu Rychlé nabíjení se nezdařilo (vypršel časový limit nebo chyba sítě). Předchozí režim byl obnoven."
-        },
-        "set_charge_mode_failed_pv_priority": {
-            "message": "Přepnutí do režimu PV Priorita se nezdařilo (vypršel časový limit nebo chyba sítě). Předchozí režim byl obnoven."
-        },
-        "set_charge_mode_failed_pv_and_battery": {
-            "message": "Přepnutí do režimu PV a baterie se nezdařilo (vypršel časový limit nebo chyba sítě). Předchozí režim byl obnoven."
-        },
-        "set_charge_power_failed": {
-            "message": "Nepodařilo se nastavit výkon nabíjení na {value} kW. Předchozí hodnota byla obnovena."
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "title": "Intervaly dotazování",
-                "description": "Nastavte, jak často integrace dotazuje SEMS portál.",
-                "data": {
-                    "scan_interval": "Interval aktualizace v klidu (sekundy)",
-                    "scan_interval_charging": "Interval aktualizace při nabíjení (sekundy)",
-                    "plant_id": "Plant ID",
-                    "product_model": "Model zařízení"
-                },
-                "data_description": {
-                    "scan_interval": "Jak často se data stahují, když se nenabíjí (doporučeno: 60)",
-                    "scan_interval_charging": "Jak často se data stahují při aktivním nabíjení (doporučeno: 30)",
-                    "plant_id": "Nechte prázdné — detekuje se automaticky.",
-                    "product_model": "Nechte prázdné — detekuje se automaticky (např. GW11K-HCA)."
-                }
-            }
-        }
-    },
-    "config": {
-        "abort": {
-            "already_configured": "Zařízení je již nakonfigurováno"
-        },
-        "error": {
-            "cannot_connect": "Nepodařilo se připojit k SEMS portálu",
-            "invalid_auth": "Neplatné přihlašovací údaje",
-            "unknown": "Neočekávaná chyba"
-        },
-        "step": {
-            "user": {
-                "title": "Připojení k SEMS portálu",
-                "description": "Zadejte přihlašovací údaje SEMS portálu. Wallbox bude detekován automaticky.",
-                "data": {
-                    "username": "Uživatelské jméno (e-mail)",
-                    "password": "Heslo"
-                }
-            },
-            "plant": {
-                "title": "Vyberte plant / stanici",
-                "description": "V účtu bylo nalezeno více plantů. Vyberte ten, který obsahuje váš wallbox.",
-                "data": {
-                    "plant_id": "Plant / stanice"
-                }
-            },
-            "charger": {
-                "title": "Vyberte wallbox",
-                "description": "Bylo nalezeno více nabíječek EV. Vyberte váš wallbox.",
-                "data": {
-                    "wallbox_serial_No": "Wallbox"
-                }
-            },
-            "charger_manual": {
-                "title": "Zadejte sériové číslo wallboxu",
-                "description": "Wallbox se nepodařilo detekovat automaticky. Zadejte sériové číslo ručně (najdete v aplikaci SEMS nebo na štítku zařízení).",
-                "data": {
-                    "wallbox_serial_No": "Sériové číslo wallboxu",
-                    "plant_id": "Plant ID (volitelné)"
-                }
-            },
-            "model": {
-                "title": "Model produktu",
-                "description": "Model produktu se nepodařilo detekovat automaticky. Najdete ho v aplikaci SEMS nebo na štítku zařízení (např. GW11K-HCA). Nechte prázdné pro přeskočení — lze zadat i dodatečně v nastavení integrace.",
-                "data": {
-                    "product_model": "Model produktu (např. GW11K-HCA)"
-                }
-            }
-        }
-    },
-    "error": {
-        "no_chargers_found": "V daném plantu nebyly nalezeny žádné nabíječky EV. Zadejte sériové číslo ručně."
-    },
-    "entity": {
-        "sensor": {
-            "status": {
-                "name": "Stav",
-                "state": {
-                    "charging": "Nabíjení",
-                    "standby": "Připraveno",
-                    "offline": "Offline",
-                    "unknown": "Neznámý"
-                }
-            },
-            "workstate": {
-                "name": "Stav vozidla",
-                "state": {
-                    "not_plugged_in": "Nepřipojeno",
-                    "connected": "Připojeno",
-                    "finished_charging": "Nabíjení dokončeno",
-                    "charged": "Nabito",
-                    "dash": "--",
-                    "unknown": "Neznámý"
-                }
-            },
-            "power": {
-                "name": "Výkon nabíjení"
-            },
-            "energy": {
-                "name": "Nabito celkem"
-            },
-            "set_charge_power_limit": {
-                "name": "Přidělený nabíjecí výkon"
-            },
-            "charge_duration": {
-                "name": "Doba nabíjení"
-            }
-        },
-        "select": {
-            "charge_mode": {
-                "name": "Režim nabíjení",
-                "state": {
-                    "fast": "Rychle",
-                    "pv_priority": "PV Priorita",
-                    "pv_and_battery": "PV a baterie"
-                }
-            }
-        },
-        "number": {
-            "charge_power": {
-                "name": "Limit výkonu"
-            }
-        },
-        "switch": {
-            "start_charging": {
-                "name": "Nabíjení"
-            },
-            "ensure_minimum_charging_power": {
-                "name": "Zajistit minimální nabíjecí výkon"
-            }
-        }
-    }
-}
