--- custom_components/sems-wallbox/translations/en.json
@@ -1,148 +1,0 @@
-{
-    "title": "GoodWe SEMS API for Wallbox",
-    "exceptions": {
-        "set_charge_mode_failed_fast": {
-            "message": "Switching to Fast charging mode failed (timeout or network error). The previous mode has been restored."
-        },
-        "set_charge_mode_failed_pv_priority": {
-            "message": "Switching to PV Priority mode failed (timeout or network error). The previous mode has been restored."
-        },
-        "set_charge_mode_failed_pv_and_battery": {
-            "message": "Switching to PV & Battery mode failed (timeout or network error). The previous mode has been restored."
-        },
-        "set_charge_power_failed": {
-            "message": "Failed to set charge power to {value} kW. The previous value has been restored."
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "title": "Polling intervals",
-                "description": "Configure how often the integration polls the SEMS portal.",
-                "data": {
-                    "scan_interval": "Idle update interval (seconds)",
-                    "scan_interval_charging": "Charging update interval (seconds)",
-                    "plant_id": "Plant ID",
-                    "product_model": "Product model"
-                },
-                "data_description": {
-                    "scan_interval": "How often to poll when not charging (recommended: 60)",
-                    "scan_interval_charging": "How often to poll while actively charging (recommended: 30)",
-                    "plant_id": "Leave empty — detected automatically.",
-                    "product_model": "Leave empty — detected automatically (e.g. GW11K-HCA)."
-                }
-            }
-        }
-    },
-    "config": {
-        "abort": {
-            "already_configured": "Device is already configured"
-        },
-        "error": {
-            "cannot_connect": "Failed to connect to SEMS portal",
-            "invalid_auth": "Invalid authentication",
-            "unknown": "Unexpected error"
-        },
-        "step": {
-            "user": {
-                "title": "Connect to SEMS portal",
-                "description": "Enter your SEMS portal credentials. Your wallbox will be detected automatically.",
-                "data": {
-                    "username": "Username (e-mail)",
-                    "password": "Password"
-                }
-            },
-            "plant": {
-                "title": "Select plant / station",
-                "description": "Multiple plants were found in your account. Select the one that contains your wallbox.",
-                "data": {
-                    "plant_id": "Plant / station"
-                }
-            },
-            "charger": {
-                "title": "Select wallbox",
-                "description": "Multiple EV chargers were found. Select your wallbox.",
-                "data": {
-                    "wallbox_serial_No": "Wallbox"
-                }
-            },
-            "charger_manual": {
-                "title": "Enter wallbox serial number",
-                "description": "The wallbox could not be detected automatically. Enter the serial number manually (found in the SEMS app or on the device label).",
-                "data": {
-                    "wallbox_serial_No": "Wallbox serial number",
-                    "plant_id": "Plant ID (optional)"
-                }
-            },
-            "model": {
-                "title": "Product model",
-                "description": "The product model could not be detected automatically. You can find it in the SEMS app or on the device label (e.g. GW11K-HCA). Leave empty to skip — you can enter it later in the integration options.",
-                "data": {
-                    "product_model": "Product model (e.g. GW11K-HCA)"
-                }
-            }
-        }
-    },
-    "error": {
-        "no_chargers_found": "No EV chargers were found in the selected plant. Enter the serial number manually."
-    },
-    "entity": {
-        "sensor": {
-            "status": {
-                "name": "Status",
-                "state": {
-                    "charging": "Charging",
-                    "standby": "Standby",
-                    "offline": "Offline",
-                    "unknown": "Unknown"
-                }
-            },
-            "workstate": {
-                "name": "Vehicle state",
-                "state": {
-                    "not_plugged_in": "Not plugged in",
-                    "connected": "Connected",
-                    "finished_charging": "Finished charging",
-                    "charged": "Charged",
-                    "dash": "--",
-                    "unknown": "Unknown"
-                }
-            },
-            "power": {
-                "name": "Charging power"
-            },
-            "energy": {
-                "name": "Total charged energy"
-            },
-            "set_charge_power_limit": {
-                "name": "Allocated charge power"
-            },
-            "charge_duration": {
-                "name": "Charge duration"
-            }
-        },
-        "select": {
-            "charge_mode": {
-                "name": "Charge mode",
-                "state": {
-                    "fast": "Fast",
-                    "pv_priority": "PV priority",
-                    "pv_and_battery": "PV & battery"
-                }
-            }
-        },
-        "number": {
-            "charge_power": {
-                "name": "Charge power limit"
-            }
-        },
-        "switch": {
-            "start_charging": {
-                "name": "Charging"
-            },
-            "ensure_minimum_charging_power": {
-                "name": "Ensure minimum charging power"
-            }
-        }
-    }
-}
