+++ custom_components/sems_wallbox/manifest.json
@@ -1,12 +1,12 @@
 {
-  "domain": "sems-wallbox",
-  "name": "GoodWe SEMS Wallbox",
+  "domain": "sems_wallbox",
+  "name": "GoodWe Wallbox",
   "codeowners": ["@prezervos", "@pedrodivisez"],
   "config_flow": true,
   "dependencies": [],
   "documentation": "https://github.com/prezervos/goodwe-wallbox-sems-home-assistant",
-  "iot_class": "cloud_polling",
+  "iot_class": "local_polling",
   "issue_tracker": "https://github.com/prezervos/goodwe-wallbox-sems-home-assistant/issues",
-  "requirements": [],
+  "requirements": ["pymodbus>=3.0.0"],
   "version": "1.4.0"
 }
