+++ hacs.json
@@ -1,4 +1,4 @@
 {
-  "name": "GoodWe SEMS Wallbox",
+  "name": "GoodWe Wallbox",
   "render_readme": true
 }
